+++ packages/react-i18n/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Fixed
+
+- Passing a number replacement value was returning an array instead of a string. [[#2292](https://github.com/Shopify/quilt/pull/2292)]
 
 ## 7.1.0 - 2022-05-19